# openapi.json
{
  "components": {
    "schemas": {
      "AuthUser": {
        "description": "The signed-in user, as recovered from a verified access token.",
        "properties": {
          "admin": {
            "default": false,
            "title": "Admin",
            "type": "boolean"
          },
          "avatar_url": {
            "default": "",
            "title": "Avatar Url",
            "type": "string"
          },
          "online_name": {
            "title": "Online Name",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "online_name"
        ],
        "title": "AuthUser",
        "type": "object"
      },
      "CreatePostRequest": {
        "properties": {
          "body": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "characters": {
            "items": {
              "type": "string"
            },
            "maxItems": 2,
            "title": "Characters",
            "type": "array"
          },
          "post_type": {
            "default": "자유",
            "title": "Post Type",
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "maxLength": 11,
                "minLength": 11,
                "pattern": "^[A-Za-z0-9_-]{11}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          }
        },
        "required": [
          "title",
          "body"
        ],
        "title": "CreatePostRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "LoginOut",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PostDetail": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "characters": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Characters",
            "type": "array"
          },
          "comments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/community__models__CommentOut"
            },
            "title": "Comments",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "post_type": {
            "default": "자유",
            "title": "Post Type",
            "type": "string"
          },
          "thumbs_down": {
            "title": "Thumbs Down",
            "type": "integer"
          },
          "thumbs_up": {
            "title": "Thumbs Up",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "maxLength": 11,
                "minLength": 11,
                "pattern": "^[A-Za-z0-9_-]{11}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          }
        },
        "required": [
          "id",
          "author",
          "title",
          "body",
          "thumbs_up",
          "thumbs_down",
          "created_at"
        ],
        "title": "PostDetail",
        "type": "object"
      },
      "PostListResponse": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PostSummary"
            },
            "title": "Posts",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "posts",
          "total",
          "page",
          "page_size"
        ],
        "title": "PostListResponse",
        "type": "object"
      },
      "PostSummary": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "characters": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Characters",
            "type": "array"
          },
          "comment_count": {
            "default": 0,
            "title": "Comment Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "post_type": {
            "default": "자유",
            "title": "Post Type",
            "type": "string"
          },
          "thumbs_down": {
            "title": "Thumbs Down",
            "type": "integer"
          },
          "thumbs_up": {
            "title": "Thumbs Up",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "maxLength": 11,
                "minLength": 11,
                "pattern": "^[A-Za-z0-9_-]{11}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          }
        },
        "required": [
          "id",
          "author",
          "title",
          "body",
          "thumbs_up",
          "thumbs_down",
          "created_at"
        ],
        "title": "PostSummary",
        "type": "object"
      },
...
      "ThumbRequest": {
        "properties": {
          "direction": {
            "description": "up for thumbs up, down for thumbs down",
            "title": "Direction",
            "type": "string"
          }
        },
        "required": [
          "direction"
        ],
        "title": "ThumbRequest",
        "type": "object"
      },
      "UpdatePostRequest": {
        "description": "The edit form submits all editable fields, including an explicit video removal.",
        "properties": {
          "body": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "characters": {
            "items": {
              "type": "string"
            },
            "maxItems": 2,
            "title": "Characters",
            "type": "array"
          },
          "post_type": {
            "title": "Post Type",
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "maxLength": 11,
                "minLength": 11,
                "pattern": "^[A-Za-z0-9_-]{11}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          }
        },
        "required": [
          "title",
          "body",
          "post_type",
          "characters",
          "youtube_video_id"
        ],
        "title": "UpdatePostRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "community__models__CommentOut": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "replies": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/community__models__CommentOut"
            },
            "title": "Replies",
            "type": "array"
          }
        },
        "required": [
          "id",
          "post_id",
          "parent_id",
          "author",
          "body",
          "created_at"
        ],
        "title": "CommentOut",
        "type": "object"
      },
      "community__models__CreateCommentRequest": {
        "properties": {
          "body": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "body"
        ],
        "title": "CreateCommentRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "`POST /auth/login`의 access_token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Live data from the RPCN multiplayer server for TTT2.",
    "title": "Tekken Tag Tournament 2 RPCN API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Verify an RPCN account and issue an access token.\n\nThere is no logout route: the token is stateless, so signing out is the\nclient discarding it.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "Auth"
        ]
      }
    },
    "/community/posts": {
      "get": {
        "operationId": "list_posts_community_posts_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "post_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Post Type"
            }
          },
          {
            "in": "query",
            "name": "characters",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "maxItems": 2,
              "title": "Characters",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Posts",
        "tags": [
          "community"
        ]
      },
      "post": {
        "operationId": "create_post_community_posts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Post",
        "tags": [
          "community"
        ]
      }
    },
    "/community/posts/{post_id}": {
      "delete": {
        "operationId": "delete_post_community_posts__post_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Post",
        "tags": [
          "community"
        ]
      },
      "get": {
        "operationId": "get_post_community_posts__post_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Post",
        "tags": [
          "community"
        ]
      },
      "patch": {
        "operationId": "update_post_community_posts__post_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Post",
        "tags": [
          "community"
        ]
      }
    },
    "/community/posts/{post_id}/comments": {
      "post": {
        "operationId": "create_comment_community_posts__post_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/community__models__CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Comment",
        "tags": [
          "community"
        ]
      }
    },
    "/community/posts/{post_id}/thumb": {
      "post": {
        "operationId": "thumb_post_community_posts__post_id__thumb_post",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThumbRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Thumb Post",
        "tags": [
          "community"
        ]
      }
    },
...
  }
}